Fix memory leaks with deleting firmware objects
