+++ src/software/simulation/BUILD
@@ -58,6 +58,7 @@ cc_library(
     srcs = ["simulator_robot_singleton.cpp"],
     hdrs = ["simulator_robot_singleton.h"],
     deps = [
+        ":firmware_object_deleter",
         ":simulator_robot",
         "//firmware/app/world:chicker",
         "//firmware/app/world:dribbler",
@@ -118,6 +119,7 @@ cc_library(
     srcs = ["simulator.cpp"],
     hdrs = ["simulator.h"],
     deps = [
+        ":firmware_object_deleter",
         ":simulator_ball",
         ":simulator_ball_singleton",
         ":simulator_robot",
@@ -192,5 +194,20 @@ cc_binary(
         "//software/logger",
         "//software/parameter:dynamic_parameters",
         "@boost//:program_options",
+    ],
+)
+
+cc_library(
+    name = "firmware_object_deleter",
+    hdrs = ["firmware_object_deleter.h"],
+    deps = [
+        ":simulator_robot",
+        "//firmware/app/world:chicker",
+        "//firmware/app/world:dribbler",
+        "//firmware/app/world:firmware_robot",
+        "//firmware/app/world:firmware_world",
+        "//firmware/app/world:wheel",
+        "//firmware/shared:physics",
+        "//software/logger",
     ],
 )
